TypeDefinition.SubTypeDefinition = function ... through Proxy.set
