+++ lib/api/types/createInstanceModificator.js
@@ -1,13 +1,5 @@
 'use strict';
 
-// const {
-// 	SymbolConstructorName,
-// } = require('../../constants');
-
-// so we should create function
-// which will return proper constructor
-// inherited from existent instance
-
 module.exports = function () {
 
 	const CreateInstanceModificator = function (ModificatorType, ModificatorTypePrototype, addProps) {
