+++ lib/api/types/createInstanceModificator200XthWay.js
@@ -5,12 +5,6 @@ module.exports = function () {
 	const CreateInstanceModificatorAncient200XthWay = function (ModificatorType, ModificatorTypePrototype, addProps) {
 
 		const existentInstance = this;
-		// For the 1st existentInstance in prototype chain
-		// ["Symbol(Defined Constructor Name)"] === MNEMOSYNE
-
-		// const BaseInstanceProto = Object.getPrototypeOf(this);
-		// const BaseInstanceConstructor = this.constructor;
-		// const BaseInstanceConstructorPrototype = this.constructor.prototype;
 
 		// const PreTripleSchemeClosure = function () {
 		const TripleSchemeClosure = function () {
