+++ lib/api/types/index.js
@@ -31,7 +31,8 @@ const compileNewModificatorFunctionBody
 const {
 	checkProto,
 	getTypeChecker,
-	CreationHandler
+	CreationHandler,
+	getTypeSplitPath
 } =	require('./utils');
 
 const TypeDescriptor = function (
@@ -200,15 +201,6 @@ const defineFromFunction = function (
 
 };
 
-const getTypeSplitPath = (path) => {
-	const split = path
-		// beautifull names
-		.replace(/\n|\t| /g, '')
-		.replace(/\[(\w+)\]/g, '.$1')
-		.replace(/^\./, '')
-		.split(/\.|\/|:/);
-	return split;
-};
 
 const define = function (TypeOrTypeName, ...args) {
