+++ lib/api/types/utils.js
@@ -57,9 +57,21 @@ const CreationHandler = function (constructionAnswer) {
 	return this;
 };
 
+
+const getTypeSplitPath = (path) => {
+	const split = path
+		// beautifull names
+		.replace(/\n|\t| /g, '')
+		.replace(/\[(\w+)\]/g, '.$1')
+		.replace(/^\./, '')
+		.split(/\.|\/|:/);
+	return split;
+};
+
 module.exports = {
 	getModificationConstructor,
 	checkProto,
 	getTypeChecker,
-	CreationHandler
+	CreationHandler,
+	getTypeSplitPath
 };
